finished working code to fit random forest multi-classification model and reorganized files
