+++ config.py
@@ -1,2 +1,15 @@
 ML_MODEL_GCS_BUCKET = "character-ml-model"
 ML_MODEL_FILE_NAME = "character_ml_model.pkl"
+
+questions = {
+    1: "You have multiple interests that you are passionate about exploring.",
+    2: "Your natural instinct is to apply an analytical approach as opposed to going with gut feeling.",
+    3: "Regardless of the ups and downs, if you have an approach that has proven successful, you prefer to stick to your process to weather the storm.",
+    4: "You are willing to accept a smaller role in exchange to be surrounded by prestige.",
+    5: "You love getting admiration from an audience, recognizing your accomplishment(s).",
+    6: "When you come across an alternative method, you feel a strong desire to add it to your repertoire.",
+    7: "You want to search out talented individuals to learn from and match up against.",
+    8: "When surrounded by more naturally, talented people, even though you might not be able to catch up to them, you feel a strong desire to maximize yourself.",
+    9: "When there is an individual around you who you think is not putting in sufficient effort, you want to completely crush him/her.",
+    10: "When you have a goal, you are willing to do whatever amount of studying/life-hacking needed to reach it.",
+}
